# PERF first-non-null